+++ src/SeerEditorManagerWidget.h
@@ -85,6 +85,7 @@ class SeerEditorManagerWidget : public QWidget, protected Ui::SeerEditorManagerW
         void                                            handleEvaluateVariableExpression    (int expressionid, QString expression);
         void                                            handleAddMemoryVisualizer           (QString expression);
         void                                            handleAddArrayVisualizer            (QString expression);
+        void                                            handleAddMatrixVisualizer           (QString expression);
         void                                            handleAddStructVisualizer           (QString expression);
         void                                            handleRequestAssembly               (QString address);
         void                                            handleRequestSourceAndAssembly      (QString address);
@@ -112,9 +113,10 @@ class SeerEditorManagerWidget : public QWidget, protected Ui::SeerEditorManagerW
         void                                            addVariableTrackerExpression        (QString expression);
         void                                            refreshVariableTrackerValues        ();
         void                                            evaluateVariableExpression          (int expressionid, QString expression);
-        void                                            addMemoryVisualize                  (QString expression);
-        void                                            addArrayVisualize                   (QString expression);
-        void                                            addStructVisualize                  (QString expression);
+        void                                            addMemoryVisualizer                 (QString expression);
+        void                                            addArrayVisualizer                  (QString expression);
+        void                                            addMatrixVisualizer                 (QString expression);
+        void                                            addStructVisualizer                 (QString expression);
         void                                            requestAssembly                     (QString address);
         void                                            requestSourceAndAssembly            (QString address);
         void                                            showMessage                         (QString message, int time);