+++ src/SeerEditorWidgetAssembly.h
@@ -100,9 +100,10 @@ class SeerEditorWidgetAssemblyArea : public SeerPlainTextEdit {
         void                                        disableBreakpoints                  (QString breakpoints);
         void                                        refreshBreakpointsStackFrames       ();
         void                                        runToAddress                        (QString address);
-        void                                        addMemoryVisualize                  (QString expression);
-        void                                        addArrayVisualize                   (QString expression);
-        void                                        addStructVisualize                  (QString expression);
+        void                                        addMemoryVisualizer                 (QString expression);
+        void                                        addArrayVisualizer                  (QString expression);
+        void                                        addMatrixVisualizer                 (QString expression);
+        void                                        addStructVisualizer                 (QString expression);
         void                                        requestAssembly                     (QString address);
         void                                        requestSourceAndAssembly            (QString address);
         void                                        showSearchBar                       (bool flag);